Moved App class to its own folder. Created a 'globals' folder.
Also updated PDF function so it would work. It seems the current implementation of pdf.add_page() doesn't support the 'format' keyword. The pages do seem to have exactly the same format as the images though.
