+++ Globals/settings.py
@@ -1,0 +1,1 @@
+IMG_DIR = "img"
